Use ComplexT as name directly. Revert registerCudaUtils to be in hpp file to prevent build issues
